some slight reorg and renaming
