+++ client_test/query_test.go
@@ -4,9 +4,7 @@ import (
 	"math/big"
 	"testing"
 
-	"github.com/google/go-cmp/cmp"
 	"github.com/statechannels/go-nitro/channel/state/outcome"
-	"github.com/statechannels/go-nitro/client"
 	"github.com/statechannels/go-nitro/client/engine/chainservice"
 	"github.com/statechannels/go-nitro/client/engine/messageservice"
 	"github.com/statechannels/go-nitro/client/query"
@@ -15,9 +13,9 @@ import (
 	"github.com/statechannels/go-nitro/types"
 )
 
-func TestLedgerLifecycle(t *testing.T) {
+func TestQueryLedgerChannel(t *testing.T) {
 
-	logFile := "test_ledger_lifecycle.log"
+	logFile := "test_query_ledger_channel.log"
 	truncateLog(logFile)
 	logDestination := newLogWriter(logFile)
 
@@ -55,10 +53,10 @@ func TestLedgerLifecycle(t *testing.T) {
 
 }
 
-func TestPaymentChannelLifecycle(t *testing.T) {
+func TestQueryPaymentChannel(t *testing.T) {
 
 	// Setup logging
-	logFile := "test_payment_lifecycle.log"
+	logFile := "test_query_payment_channel.log"
 	truncateLog(logFile)
 	logDestination := newLogWriter(logFile)
 
@@ -140,53 +138,4 @@ func expectedPaymentInfo(id types.Destination, outcome outcome.Exit, status quer
 			RemainingFunds: outcome[0].Allocations[0].Amount,
 			PaidSoFar:      outcome[0].Allocations[1].Amount,
 		}}
-}
-
-// checkPaymentChannel checks that the ledger channel has the expected outcome and status
-// It will fail if the channel does not exist
-func checkPaymentChannel(t *testing.T, id types.Destination, o outcome.Exit, status query.ChannelStatus, clients ...*client.Client) {
-
-	for _, c := range clients {
-		expected := expectedPaymentInfo(id, o, status)
-		ledger, err := c.GetPaymentChannel(id)
-		if err != nil {
-			t.Fatal(err)
-		}
-		if diff := cmp.Diff(expected, ledger, cmp.AllowUnexported(big.Int{})); diff != "" {
-			t.Fatalf("Payment channel diff mismatch (-want +got):\n%s", diff)
-		}
-	}
-}
-
-// expectedLedgerInfo constructs a LedgerChannelInfo so we can easily compare it to the result of GetLedgerChannel
-func expectedLedgerInfo(id types.Destination, outcome outcome.Exit, status query.ChannelStatus) query.LedgerChannelInfo {
-	clientAdd, _ := outcome[0].Allocations[0].Destination.ToAddress()
-	hubAdd, _ := outcome[0].Allocations[1].Destination.ToAddress()
-
-	return query.LedgerChannelInfo{
-		ID:     id,
-		Status: status,
-		Balance: query.LedgerChannelBalance{
-			AssetAddress:  types.Address{},
-			Hub:           hubAdd,
-			Client:        clientAdd,
-			ClientBalance: outcome[0].Allocations[0].Amount,
-			HubBalance:    outcome[0].Allocations[1].Amount,
-		}}
-}
-
-// checkLedgerChannel checks that the ledger channel has the expected outcome and status
-// It will fail if the channel does not exist
-func checkLedgerChannel(t *testing.T, ledgerId types.Destination, o outcome.Exit, status query.ChannelStatus, clients ...*client.Client) {
-
-	for _, c := range clients {
-		expected := expectedLedgerInfo(ledgerId, o, status)
-		ledger, err := c.GetLedgerChannel(ledgerId)
-		if err != nil {
-			t.Fatal(err)
-		}
-		if diff := cmp.Diff(expected, ledger, cmp.AllowUnexported(big.Int{})); diff != "" {
-			t.Fatalf("Ledger diff mismatch (-want +got):\n%s", diff)
-		}
-	}
 }
